Added physical meaning to the linear correction
Kept the name "tilt_x" and "tilt_y" but it would be better to change them to "shift".
Now the user enters a shift X/Y in mm to move the image in the Fourier plane.

Need to be done : the focal length of the post SLM used is needed for the calculation but for now we store it only for the gbsax algo in the toml file. Myabe it would be good to use a general parameter ? LINE :  focal_postSLM = Q_(self._plugin_config('algo')['gbsax']['focal_length_mm'], 'mm')